+++ crates/nostr/src/event/unsigned.rs
@@ -53,7 +53,7 @@ impl UnsignedEvent {
             pubkey: public_key,
             created_at,
             kind,
-            tags: Tags::new(tags.into_iter().collect()),
+            tags: Tags::from_list(tags.into_iter().collect()),
             content: content.into(),
         }
     }